api/requirements.txt
# Restful API
djangorestframework
markdown
django-filter
